untitled folder/website-master/requirements.txt
alphabet-detector==0.0.5
appdirs==1.4.2
BeautifulSoup==3.2.1
boto==2.43.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
coreapi==2.3.1
coreschema==0.0.4
decorator==4.0.11
defusedxml==0.4.1
dj-database-url==0.4.0
Django==1.11.3
django-allauth==0.32.0
django-appconf==1.0.2
django-autoslug==1.9.3
django-cities-light==3.3.0
django-contrib-comments==1.7.3
django-dynamic-scraper==0.11.4
django-filter==1.0.4
django-geoposition==0.3.0
django-imagekit==3.3
django-model-utils==2.6
django-mptt==0.8.6
django-postgrespool==0.3.0
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
django-rosetta==0.7.12
django-storages==1.5.1
django-tinymce4-lite==1.4.1
django-updown==1.0.1
django-user-agents==0.3.0
djangorestframework==3.6.3
djangorestframework-gis==0.11
drfdocs==0.0.11
enum34==1.1.6
firebase-token-generator==2.0.1
future==0.15.2
futures==3.0.5
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
geopy==1.11.0
google-gax==0.15.0
googleapis-common-protos==1.5.0
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.0.1
gunicorn==19.2.1
httplib2==0.9.2
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
jsonpath-rw==1.4.0
jws==0.1.3
Markdown==2.6.1
MarkupSafe==1.0
microsofttranslator==0.8
oauth2client==3.0.0
oauthlib==2.0.2
openapi-codec==1.3.1
packaging==16.8
pilkit==1.1.13
Pillow==3.4.2
ply==3.8
polib==1.0.8
progressbar2==3.12.0
protobuf==3.1.0.post1
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycryptodome==3.4.3
pyparsing==2.2.0
Pyrebase==3.0.24
python-dotenv==0.6.2
-e git://github.com/Sosokan/python-firebase@5d63d8e03fabacc270c5bfae73e4bf5b50fc3d51#egg=python_firebase
python-jwt==2.0.1
python-openid==2.2.5
python-utils==2.0.1
python3-openid==3.0.10
pytz==2017.2
qrcode==5.3
queuelib==1.4.2
requests==2.18.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.0
rollbar==0.13.11
rsa==3.4.2
simplejson==3.10.0
six==1.10.0
sorl-thumbnail==12.4a1
sorl-thumbnail-serializer-field==0.1
SQLAlchemy==0.9.8
swapper==1.0.0
tqdm==4.10.0
ua-parser==0.7.2
Unidecode==0.4.19
uritemplate==3.0.0
urllib3==1.21.1
user-agents==1.0.1
whitenoise==3.2.2
django-pylibmc==0.6.1
pylibmc==1.5.1
